Согласованные имена команд

+++ servers/relay/Server/src/room/command/float.rs
@@ -20,7 +20,7 @@ impl ServerCommandExecutor for IncrementFloat64C2SCommand {
 			let access_groups = object.access_groups.clone();
 			room.send_to_group(
 				access_groups,
-				S2CCommand::SetFloat64(SetFloat64Command {
+				S2CCommand::SetFloat(SetFloat64Command {
 					object_id: self.object_id,
 					field_id: self.field_id,
 					value,
@@ -35,7 +35,7 @@ impl ServerCommandExecutor for SetFloat64Command {
 		if let Some(object) = room.get_object_mut(&self.object_id) {
 			object.floats.insert(self.field_id, self.value);
 			let access_groups = object.access_groups;
-			room.send_to_group(access_groups, S2CCommand::SetFloat64(self));
+			room.send_to_group(access_groups, S2CCommand::SetFloat(self));
 		}
 	}
 }
@@ -43,7 +43,7 @@ impl ServerCommandExecutor for SetFloat64Command {
 impl GameObject {
 	pub fn floats_to_commands(&self, commands: &mut Vec<S2CCommand>) {
 		self.floats.iter().for_each(|(k, v)| {
-			commands.push(S2CCommand::SetFloat64(SetFloat64Command {
+			commands.push(S2CCommand::SetFloat(SetFloat64Command {
 				object_id: self.id.clone(),
 				field_id: k.clone(),
 				value: *v,
@@ -60,7 +60,7 @@ mod tests {
 	use cheetah_relay_common::room::owner::ObjectOwner;
 
 	use crate::room::command::ServerCommandExecutor;
-	
+
 	use crate::room::Room;
 
 	#[test]
@@ -77,7 +77,7 @@ mod tests {
 
 		let object = room.get_object_mut(&object_id).unwrap();
 		assert_eq!(*object.floats.get(&10).unwrap() as u64, 100);
-		assert!(matches!(room.out_commands.pop_back(), Some((.., S2CCommand::SetFloat64(c))) if c==command));
+		assert!(matches!(room.out_commands.pop_back(), Some((.., S2CCommand::SetFloat(c))) if c==command));
 	}
 
 	#[test]
@@ -102,7 +102,7 @@ mod tests {
 			value: 200.200,
 		};
 		room.out_commands.pop_back();
-		assert!(matches!(room.out_commands.pop_back(), Some((.., S2CCommand::SetFloat64(c))) if c==result));
+		assert!(matches!(room.out_commands.pop_back(), Some((.., S2CCommand::SetFloat(c))) if c==result));
 	}
 
 	#[test]
